backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
httptools==0.6.4
idna==3.10
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
pycryptodomex==3.23.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tiktoken==0.10.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yt-dlp==2025.7.21